requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.21
pydantic==1.10.12
passlib[bcrypt]==1.7.4
psycopg2-binary
alembic==1.10.1
python-jose
python-multipart